feat(home): add new home page with navigation cards for calculator and management
- Introduce a new HomePage component as the landing page
- Provide two main navigation cards: Steel Calculator and Construction Management
- Each card includes icon, title, description, and button to navigate to respective pages
- Remove previous complex calculator and project selection UI from root page
- Simplify imports and reduce bundle size by removing unused components
- Enhance user experience with clear workspace choice on app start
