+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const c=(c,i)=>(c=new URL(c+".js",i).href,s[c]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()}).then(()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e}));self.define=(i,r)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let t={};const a=e=>c(e,n),d={module:{uri:n},exports:t,require:a};s[n]=Promise.all(i.map(e=>d[e]||a(e))).then(e=>(r(...e),t))}}define(["./workbox-cb477421"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"71411bd2816d7704f2ea25dc976385d3"},{url:"/_next/static/chunks/172-c30f98b2a1fde2d7.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/218.1c7b374dc34d4f05.js",revision:"1c7b374dc34d4f05"},{url:"/_next/static/chunks/376.f847533223f57832.js",revision:"f847533223f57832"},{url:"/_next/static/chunks/445.89554df3120ae451.js",revision:"89554df3120ae451"},{url:"/_next/static/chunks/4bd1b696-c4194431a1dff530.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/510.1dc8ba195f048df0.js",revision:"1dc8ba195f048df0"},{url:"/_next/static/chunks/514-0904908432fdf56f.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/684-9b408d877f485efc.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/772.67fb89c7e9a32689.js",revision:"67fb89c7e9a32689"},{url:"/_next/static/chunks/815-4404c3bf7b979f1a.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/917-d11c2a839fff2f33.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/918.b0b061b5b8e35075.js",revision:"b0b061b5b8e35075"},{url:"/_next/static/chunks/933.add5c3788b063b31.js",revision:"add5c3788b063b31"},{url:"/_next/static/chunks/app/_not-found/page-231fc603c74735bf.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/app/layout-77d92da6035d2e40.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/app/page-7bee07df34bd0dc1.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/framework-6d868e9bc95e10d8.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/main-73916da36e825e86.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/main-app-f5531582a8aef103.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/pages/_app-da15c11dea942c36.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/pages/_error-cc3f077a18ea1793.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-809dcbc79bf4aa01.js",revision:"qS1rCprbwFCwpR5_TjTgR"},{url:"/_next/static/css/043fbdebc03d8114.css",revision:"043fbdebc03d8114"},{url:"/_next/static/css/931b179cff3131dc.css",revision:"931b179cff3131dc"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/_next/static/qS1rCprbwFCwpR5_TjTgR/_buildManifest.js",revision:"e7ac8479043ecfe785d0a4c7e19745bb"},{url:"/_next/static/qS1rCprbwFCwpR5_TjTgR/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/apple-icon.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/icon-192x192.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/icon-512x512.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/manifest.json",revision:"1aef340bf0044412798c4c92218d5c81"},{url:"/placeholder-logo.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/placeholder-logo.svg",revision:"1e16dc7df824652c5906a2ab44aef78c"},{url:"/placeholder-user.jpg",revision:"82c9573f1276f9683ba7d92d8a8c6edd"},{url:"/placeholder.jpg",revision:"887632fd67dd19a0d58abde79d8e2640"},{url:"/placeholder.svg",revision:"35707bd9960ba5281c72af927b79291f"},{url:"/sw-custom.js",revision:"116aa56ef5fbbb60409ec60f02a53631"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https?.*/,new e.NetworkFirst({cacheName:"offlineCache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:2592e3})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0});
+if(!self.define){let e,s={};const c=(c,i)=>(c=new URL(c+".js",i).href,s[c]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()}).then(()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e}));self.define=(i,n)=>{const r=e||("document"in self?document.currentScript.src:"")||location.href;if(s[r])return;let t={};const a=e=>c(e,r),d={module:{uri:r},exports:t,require:a};s[r]=Promise.all(i.map(e=>d[e]||a(e))).then(e=>(n(...e),t))}}define(["./workbox-cb477421"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"3b3e74532f542b09ad58d3327cbd7309"},{url:"/_next/static/chunks/186-9e3b58a7df50afba.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/263-dd7f9b5c3f0beee8.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/266-9d0b1659b7dc2466.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/445.89554df3120ae451.js",revision:"89554df3120ae451"},{url:"/_next/static/chunks/4bd1b696-c4194431a1dff530.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/523-96b706f7960c9ca9.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/684-9b408d877f485efc.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/695-320ac93398c3f778.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/777-08354a411ca3ddf2.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/880-55b62d1c136f9d4b.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/933.add5c3788b063b31.js",revision:"add5c3788b063b31"},{url:"/_next/static/chunks/991-9d52be6dea676934.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/app/_not-found/page-231fc603c74735bf.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/app/calculator/page-bb74e3007dc5143b.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/app/layout-2d724ea411fc1121.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/app/management/page-3aebcc7ee8ed64ed.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/app/page-0cf62aca13109b60.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/framework-6d868e9bc95e10d8.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/main-73916da36e825e86.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/main-app-4815a13759d217f6.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/pages/_app-da15c11dea942c36.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/pages/_error-cc3f077a18ea1793.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-c73d29cab829cc56.js",revision:"hy2YGj_G1_vGKVr2wmzFs"},{url:"/_next/static/css/12cfa661d0fc709e.css",revision:"12cfa661d0fc709e"},{url:"/_next/static/css/931b179cff3131dc.css",revision:"931b179cff3131dc"},{url:"/_next/static/hy2YGj_G1_vGKVr2wmzFs/_buildManifest.js",revision:"b70f8b132274f639bbf79b654e2bbd35"},{url:"/_next/static/hy2YGj_G1_vGKVr2wmzFs/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/apple-icon.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/icon-192x192.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/icon-512x512.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/manifest.json",revision:"1aef340bf0044412798c4c92218d5c81"},{url:"/placeholder-logo.png",revision:"b7d4c7dd55cf683c956391f9c2ce3f5b"},{url:"/placeholder-logo.svg",revision:"1e16dc7df824652c5906a2ab44aef78c"},{url:"/placeholder-user.jpg",revision:"82c9573f1276f9683ba7d92d8a8c6edd"},{url:"/placeholder.jpg",revision:"887632fd67dd19a0d58abde79d8e2640"},{url:"/placeholder.svg",revision:"35707bd9960ba5281c72af927b79291f"},{url:"/sw-custom.js",revision:"116aa56ef5fbbb60409ec60f02a53631"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https?.*/,new e.NetworkFirst({cacheName:"offlineCache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:2592e3})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0});
